LayoutTests/fast/scrolling/ios/overflow-scroll-overlap-4.html

<!DOCTYPE html> <!-- webkit-test-runner [ internal:AsyncOverflowScrollingEnabled=true ] -->
<html>
<head>
<meta name="viewport" content="width=device-width, initial-scale=1">
<script src="../../../resources/ui-helper.js"></script>
<script src="../resources/overflow-scroll-overlap.js"></script>
<style>
.case {
    width: 200px;
    height: 200px;
    display: inline-block;
    position: relative;
}
.scrollcontent {
    width: 500px;
    height: 500px;
    background: green;
}

.overflowscroll {
    overflow: scroll;
    height: 100px;
    width: 100px;
    position: absolute;
    border: 2px solid black;
}
.overlapping {
    position:absolute;
    left: 25px;
    top: 25px;
    width: 100px;
    height: 100px;
    background: red;
}
.clip {
    position:absolute;
    width: 100px;
    height: 100px;
    overflow:hidden;
}
.large {
    width: 3000px;
    height: 150px;
}
#log {
    position:relative;
    white-space: pre;
}
</style>
</head>
<body onload="runTest()">
<p>
Test that scrollable areas with overflowing overlap are correctly targeted.
</p>

<div class="case">
    <div class="overflowscroll target">
        <div class="scrollcontent"></div>
    </div>
    <div class="overlapping" style="left: 80px">
        <div class="overlapping" style="left: -40px; top: -30px"></div>
    </div>
</div>

<div class="case">
    <div class="overflowscroll target">
        <div class="scrollcontent"></div>
    </div>
    <div class="overlapping" style="left: 80px; z-index:0">
        <div class="overlapping" style="left: -40px; top: -30px"></div>
    </div>
</div>

<div class="case">
    <div class="overflowscroll target">
        <div class="scrollcontent"></div>
    </div>
    <div class="overlapping" style="left: 80px; top: 90px; z-index:0">
        <div class="overlapping" style="left: -40px; top: -120px"></div>
    </div>
</div>

<div class="case">
    <div class="overflowscroll target">
        <div class="scrollcontent"></div>
    </div>
    <div class="overlapping" style="left: 80px; top: 90px; z-index:0">
        <div class="overlapping" style="left: -40px; top: -120px; z-index:-1"></div>
    </div>
</div>

<div class="case">
    <div class="overflowscroll target">
        <div class="scrollcontent"></div>
    </div>
    <div class="overlapping" style="left: 80px; top: 90px; z-index:0">
        <div class="overlapping" style="left: -40px; top: -120px; z-index:1"></div>
    </div>
</div>

<div class="case">
    <div class="overflowscroll target">
        <div class="scrollcontent"></div>
    </div>
    <div class="overlapping" style="left: 80px; top: 60px; z-index:0">
        <div class="overlapping" style="left: -40px; top: -120px"></div>
    </div>
</div>

<div id=log></div>

</body>
</html>
